Finalize PipeAndFilter2
Finished the implimentation for PipeAndFilter2, with this
implimentation each filter is an independent java class that
can be run independely, each taking in a text input, and
its output is a string that can be used as the input to
the next java class. The filters in this pipe and filter
design are implimented as the jvm environment as each program
output is piped as input into the next. From the command line
you can run this to execute the full program when Java.exe is
in your PATH:

Java getInputFile | Java fileReader| Java circularShifter |
Java sorter | Java output